Wrapper for prepare_data_normalization avg_ranking and calculate_norm_ranks

Wrapper for prepare_data_normalization avg_ranking and calculate_norm_ranks

Usage

get_final_ranking_values(
  sum_exp,
  assay_to_use,
  stable_genes = NULL,
  most_variable_genes = "small",
  verbose = 1
)

Arguments

sum_exp

Summarized experiment object. It should contain a "rank" slot that will be used to calculate the average rank of the stable genes.

assay_to_use

A string. Which assay to use when calculating the average expression of the stable genes.

stable_genes

A vector. A list with 44 genes that will be used to the qPCR-like normalization if NULL (default).

most_variable_genes

A vector. Most variable genes determined by some procedure. Default is "small" and uses an internal list of 1000 genes to calculate the embedding. If using for calculating the regressed data, use the "big" option.

Value

A summarized experiment object containing the normalized data that can be applied later on to obtain the embedding using the loadings of the PCA.

Examples

if (FALSE) { # \dontrun{
get_final_ranking_values(tcga, "logFPKM_TMM")
} # }